==> 1-Dataset.ipynb <==
# Dataset

from google.cloud import bigquery
import google.auth
import csv
import time
from tqdm.notebook import tqdm
import sys
import sklearn.model_selection
import pandas as pd
import hashlib
import random
import os

csv.field_size_limit(sys.maxsize) # to avoid error: _csv.Error: field larger than field limit (131072)

dataset_all_file = 'nlbse23-issue-classification-all.csv'
dataset_file = 'nlbse23-issue-classification.csv'
train_file = 'nlbse23-issue-classification-train.csv'
test_file = 'nlbse23-issue-classification-test.csv'



def count_csv_rows(filename):
	with open(filename, "r", newline='', encoding='utf-8') as f:
		return sum(1 for _ in csv.DictReader(f))

def print_csv_preview(filename):
	print(filename)
	print("total rows", count_csv_rows(filename))
	display(pd.read_csv(filename, nrows=5))

def print_label_stats(filename):
	label_count = {"bug": 0, "feature": 0, "question": 0, "documentation": 0}
	with open(filename, "r", newline='', encoding='utf-8') as f:
		for row in tqdm(csv.DictReader(f), desc="Counting labels", total=count_csv_rows(filename)):
			label_count[row["labels"]] += 1
	print(filename)
	print("label counts", label_count)
	total = sum(label_count.values())
	print("label distribution", {k: round(v / total, 3) for k, v in label_count.items()})



# https://cloud.google.com/docs/authentication/application-default-credentials
client = bigquery.Client()

query = """
WITH
  label_synonyms AS (
    SELECT 'bug' AS label, synonym FROM UNNEST(['bug', 'type: bug', 'kind/bug', 'crash', 'defect', 'type-defect', 'type:bug', 'browser bug', 'fix', 'fixed', 'bugfix', 'bug fix', 'resolution: fixed', 'troubleshooting', 'type/bug', 'bug report']) AS synonym
    UNION ALL
    SELECT 'feature', * FROM UNNEST(['feature', 'feature request', 'enhancement', 'improvement', 'type: feature', 'type:feature', 'new feature', 'kind/feature', 'kind/enhancement'])
    UNION ALL
    SELECT 'question', * FROM UNNEST(['question', 'faq', 'type: question', 'type:question'])
    UNION ALL
    SELECT 'documentation', * FROM UNNEST(['documentation', 'docs', 'doc', 'type: documentation', 'needs documentation', 'area/documentation', 'type: docs', 'type:docs', 'needs docs', 'wiki', 'kind/documentation', 'kind/docs'])
  ),
  close_events AS (
    SELECT payload
    FROM `githubarchive.day.2022*`
    WHERE
      _TABLE_SUFFIX BETWEEN '0101' AND '0101' -- FREE
      -- _TABLE_SUFFIX BETWEEN '0101' AND '0131' -- $
      -- _TABLE_SUFFIX BETWEEN '0101' AND '0930' -- $$$ $$$ $$$
      AND type = 'IssuesEvent'
      AND JSON_EXTRACT_SCALAR(payload, '$.action') = 'closed'
      AND JSON_EXTRACT_SCALAR(payload, '$.issue.body') != 'null'
  ),
  nested_labels AS (
    SELECT
      ARRAY(
        SELECT LOWER(JSON_EXTRACT_SCALAR(label_payload, '$.name'))
        FROM UNNEST(JSON_EXTRACT_ARRAY(payload, '$.issue.labels')) AS label_payload
      ) AS labels,
      payload
    FROM close_events
  ),
  synonymized_labels AS (
    SELECT
      ARRAY(
        SELECT DISTINCT label_synonyms.label -- bug, feature, question, documentation
        FROM UNNEST(nested_labels.labels) AS label, label_synonyms
        WHERE label = label_synonyms.synonym
        ORDER BY label_synonyms.label
      ) AS labels,
      payload
    FROM nested_labels
  ),
  filtered_labels AS (
    SELECT labels, payload
    FROM synonymized_labels
    WHERE ARRAY_LENGTH(labels) = 1 -- exactly 1 label
  ),
  concatenated_labels AS (
    SELECT ARRAY_TO_STRING(labels, ',') AS labels, payload
    FROM filtered_labels
  )
SELECT
  JSON_EXTRACT_SCALAR(payload, '$.issue.id') AS id,
  labels,
  JSON_EXTRACT_SCALAR(payload, '$.issue.title') AS title,
  JSON_EXTRACT_SCALAR(payload, '$.issue.body') AS body,
  JSON_EXTRACT_SCALAR(payload, '$.issue.author_association') AS author_association
FROM concatenated_labels
"""

query_job = client.query(query)
rows = query_job.result()

with open(dataset_all_file, "w", newline='', encoding='utf-8') as f:
  fieldnames = ["id", "labels", "title", "body", "author_association"]
  writer = csv.DictWriter(f, fieldnames=fieldnames)
  writer.writeheader()
  for row in tqdm(rows, desc="BigQuery", smoothing=0):
    writer.writerow({
      **row,
      "title": row["title"].replace("\0", ""),
      "body": row["body"].replace("\0", ""),
    })
    
print_csv_preview(dataset_all_file)

print_label_stats(dataset_all_file)

# Language

We filter the dataset to only include English issues.

import fasttext
import gensim

# https://fasttext.cc/docs/en/language-identification.html
# download the training set if it does not exist
if not os.path.isfile("lid.176.bin"):
  !curl "https://dl.fbaipublicfiles.com/fasttext/supervised-models/lid.176.bin" -o "lid.176.bin"

model = fasttext.load_model("lid.176.bin")

def identify_language(text):
  text = gensim.parsing.strip_multiple_whitespaces(text)
  return model.predict(text)[0][0][9:]

def print_language_stats(filename):
	# compute language statistsics
	language_count = {}
	with open(filename, "r", newline='', encoding='utf-8') as f:
		rows = csv.DictReader(f)
		texts = (f"{row['title']} {row['body']}" for row in rows)
		langs = (identify_language(text) for text in texts)
		for lang in tqdm(langs, desc="Language stats", total=count_csv_rows(filename)):
			if lang not in language_count:
				language_count[lang] = 0
			language_count[lang] += 1
	total = sum(language_count.values())
	print(filename)
	print("language counts", language_count)
	print("language distribution", {k: round(v / total, 3) for k, v in language_count.items()})

print_language_stats(dataset_all_file)

def filter_en(in_file, out_file):
	with open(in_file, "r", newline='', encoding='utf-8') as f_in, open(out_file, "w", newline='', encoding='utf-8') as f_out:
		reader = csv.DictReader(f_in)
		writer = csv.DictWriter(f_out, fieldnames=reader.fieldnames)
		writer.writeheader()
		for row in tqdm(reader, desc="Filtering EN", total=count_csv_rows(filename_csv_in)):
			if identify_language(f"{row['title']} {row['body']}") == "en":
				writer.writerow(row)

filter_en(dataset_all_file, dataset_file)

print_csv_preview(dataset_file)
print_label_stats(dataset_file)

# Training and Test Split

We split the dataset into training and test sets.

def split_dataset(in_file, out_train_file, out_test_file, train_percentage):
	seed = int(hashlib.sha256("nlbse2023".encode('utf8')).hexdigest(), 16)
	r = random.Random(seed)

	with open(in_file, "r", newline='') as f_in, open(out_train_file, "w", newline='') as f_train, open(out_test_file, "w", newline='') as f_test:
		reader = csv.DictReader(f_in)
		writer_train = csv.DictWriter(f_train, fieldnames=reader.fieldnames)
		writer_train.writeheader()
		writer_test = csv.DictWriter(f_test, fieldnames=reader.fieldnames)
		writer_test.writeheader()
		total = count_csv_rows(in_file)
		is_train_gen = (r.random() < 0.9 for _ in range(total))
		for row, is_train in tqdm(zip(reader, is_train_gen), desc="Splitting", total=total, smoothing=0):
			if is_train:
				writer_train.writerow(row)
			else:
				writer_test.writerow(row)

split_dataset(dataset_file, train_file, test_file, train_percentage=0.9)

print_csv_preview(train_file)
print_label_stats(train_file)

print_csv_preview(test_file)
print_label_stats(test_file)

!tar -czf {train_file}.tar.gz {train_file}
!tar -czf {test_file}.tar.gz {test_file}
