==> 2-Template-Transformers.ipynb <==
import os
import time
import numpy as np
import pandas as pd
import seaborn as sns
import gensim
import sklearn.metrics
import re
import unicodedata
from tqdm.notebook import tqdm
import torch
from simpletransformers.classification import ClassificationModel, ClassificationArgs
import wandb
import csv
from functools import partial
import logging

tqdm.pandas()

logging.basicConfig(level=logging.INFO)
transformers_logger = logging.getLogger("transformers")
transformers_logger.setLevel(logging.WARNING)

# download the training set if it does not exist
if not os.path.isfile("github-labels-top3-803k-train.csv"):
  !curl "https://tickettagger.blob.core.windows.net/datasets/github-labels-top3-803k-train.tar.gz" | tar -xz

train_df = pd.read_csv("github-labels-top3-803k-train.csv")
display(train_df.head(5))

if not os.path.isfile("github-labels-top3-803k-test.csv"):
  !curl "https://tickettagger.blob.core.windows.net/datasets/github-labels-top3-803k-test.tar.gz" | tar -xz

test_df = pd.read_csv("github-labels-top3-803k-test.csv")
display(test_df.head(5))


function_sig_regex = re.compile(r'[a-zA-Z][a-zA-Z0-9_.]*\([a-zA-Z0-9_, ]*\)')
issue_id_regex = re.compile(r'#[0-9]+')
non_ascii_char_regex = re.compile(r'[^\x00-\x7f]')

def count_tokens(text):
	return text.count(" ") + 1

def preprocess(text, max_tokens=None):
  text = str(text)

  # lowercase
  text = text.lower()

  # replace function signatures
  text = function_sig_regex.sub(" function ", text)

  # replace issue ids
  text = issue_id_regex.sub(" issue ", text)
  
  # remove html tags
  text = gensim.parsing.preprocessing.strip_tags(text)
  
  # remove punctuation
  text = gensim.parsing.preprocessing.strip_punctuation(text)
  
  # remove numerics
  text = gensim.parsing.preprocessing.strip_numeric(text)
  
  # remove non-ascii characters
  text = non_ascii_char_regex.sub("", text)
  
  text = unicodedata.normalize('NFD', text)
  
  # remove consecutive whitespace characters and convert tabs to spaces
  text = gensim.parsing.preprocessing.strip_multiple_whitespaces(text)
  
  text = gensim.parsing.preprocessing.strip_short(text, minsize=3)
  
  text = gensim.parsing.preprocessing.remove_stopwords(text)
  
  # text = gensim.parsing.preprocessing.stem_text(text)
  
  # limit the number of tokens
  if max_tokens is not None:
    text = " ".join(text.split()[:max_tokens])
  
  return text

q=[.5, .75, .8, .85, .9, .95, .99, .999]

display(
	"title token frequency quantiles", 
	train_df["issue_title"].sample(10_000).progress_apply(preprocess).apply(count_tokens).quantile(q=q)
)
display(
	"body token frequency quantiles", 
	train_df["issue_body"].sample(10_000).progress_apply(preprocess).apply(count_tokens).quantile(q=q)
)

def preprocess_row(row):
  doc = "author " + row["issue_author_association"].lower()
  doc += " title " + preprocess(row["issue_title"], max_tokens=18) # 99.9% of titles have <= 18 tokens
  doc += " body " + preprocess(row["issue_body"], max_tokens=511-count_tokens(doc))

  assert count_tokens(doc) <= 512

  return doc

# transform dataset into simpletransformers format
# https://simpletransformers.ai/docs/classification-data-formats/#multi-class-classification

# train_df["text"] = train_df.progress_apply(preprocess_row, axis=1)
# train_df["labels"] = pd.Categorical(train_df["issue_label"]).codes
# display(train_df[["text", "labels"]].head(5))
# train_df[["text", "labels"]].to_csv("train.csv", index=False)

# test_df["text"] = test_df.progress_apply(preprocess_row, axis=1)
# test_df["labels"] = pd.Categorical(test_df["issue_label"]).codes
# display(test_df[["text", "labels"]].head(5))
# test_df[["text", "labels"]].to_csv("test.csv", index=False)

def transform_to_simpletransformers_format(i_path, o_path):
	label_map = {"bug": 0, "enhancement": 1, "question": 2}
	n_lines = sum(1 for _ in open(i_path, "r"))

	with open(i_path, "r") as i_f, open(o_path, "w") as o_f:
		reader = csv.DictReader(i_f)
		writer = csv.DictWriter(o_f, fieldnames=["text", "labels"])
		writer.writeheader()
		for row in tqdm(reader, desc="Transform to simpletransformers format", total=n_lines):
			text = preprocess_row(row)
			labels = label_map[row["issue_label"]]
			writer.writerow({"text": text, "labels": labels})

transform_to_simpletransformers_format("github-labels-top3-803k-train.csv", "train.csv")
transform_to_simpletransformers_format("github-labels-top3-803k-test.csv", "test.csv")

model_args = ClassificationArgs()

# https://simpletransformers.ai/docs/classification-specifics/#lazy-loading-data
model_args.lazy_loading = True
model_args.lazy_delimiter = ','

# 
# https://github.com/ThilinaRajapakse/simpletransformers/issues/225
# model_args.use_multiprocessing = False
# model_args.use_multiprocessing_for_evaluation = False
# model_args.multiprocessing_chunksize = 1
# model_args.dataloader_num_workers = 1

# model_args.learning_rate = 3e-5 # 4e-5
model_args.num_train_epochs = 4 # 1

# ~20 for title, ~100 for body
model_args.max_seq_length = 128

# batch_size = 32 # 5.2GB VRAM
# batch_size = 64 # 7.3GB VRAM
# batch_size = 96 # 9.4GB VRAM
batch_size = 128 # 12GB VRAM
model_args.train_batch_size = batch_size
model_args.eval_batch_size = batch_size

model_args.save_steps = -1
# model_args.save_model_every_epoch = False

# miscallenous
model_args.manual_seed = 0
model_args.evaluate_during_training = True
model_args.overwrite_output_dir = True
model_args.wandb_project = "NLBSE 2023 Template"

model = ClassificationModel(
  'roberta', 
  'roberta-base', 
  args=model_args, 
  num_labels=3
)

metrics = {
  "p_micro": partial(sklearn.metrics.precision_score, average='micro'),
  "r_micro": partial(sklearn.metrics.recall_score, average='micro'),
  "f1_micro": partial(sklearn.metrics.f1_score, average='micro'),
}

# model.train_model(train_df=train_df, eval_df=test_df, **metrics)
model.train_model(train_df="train.csv", eval_df="test.csv", **metrics)
