==> 3-Template-RoBERTa.ipynb <==
!nvidia-smi

import os
import time
import numpy as np
import pandas as pd
import gensim
import sklearn.metrics
import re
import unicodedata
from tqdm import tqdm
import torch
from simpletransformers.classification import ClassificationModel, ClassificationArgs
import wandb
import csv
from functools import partial
import itertools
import random
import sys
import hashlib
import time

train_file = 'nlbse23-issue-classification-train.csv'
test_file = 'nlbse23-issue-classification-test.csv'

csv.field_size_limit(sys.maxsize) # to avoid error: _csv.Error: field larger than field limit (131072)

def count_tokens(text):
	return len(text.split())

def count_csv_rows(csv_file):
	with open(csv_file, 'r', newline='', encoding='utf-8') as f:
		return sum(1 for _ in csv.DictReader(f))

def print_csv_preview(filename, sep=None):
	print(filename)
	print("total rows", count_csv_rows(filename))
	display(pd.read_csv(filename, nrows=5, sep=sep))

def sample_csv(file, n_sample):
	n_population = count_csv_rows(file)
	skiprows = random.sample(range(1, n_population), n_population - n_sample)
	return pd.read_csv(file, skiprows=skiprows)


### Download dataset

# download the training set if it does not exist
if not os.path.isfile(train_file):
  !curl "https://tickettagger.blob.core.windows.net/datasets/{train_file}.tar.gz" | tar -xz

print_csv_preview(train_file)

if not os.path.isfile(test_file):
  !curl "https://tickettagger.blob.core.windows.net/datasets/{test_file}.tar.gz" | tar -xz

print_csv_preview(test_file)

### Preprocess

function_sig_regex = re.compile(r'[a-zA-Z][a-zA-Z0-9_.]*\([a-zA-Z0-9_, ]*\)')
issue_id_regex = re.compile(r'#[0-9]+')
non_ascii_char_regex = re.compile(r'[^\x00-\x7f]')
punctuations = '!\'"`$%&\()*,/:;<=>[\\]^{|}~+#@-_'
punctuations_trans = str.maketrans(punctuations, " " * len(punctuations))

def preprocess(text, max_tokens=None):
  text = str(text)

  # replace function signatures
  text = function_sig_regex.sub(" FUNCTION ", text)

  # replace issue ids
  text = issue_id_regex.sub(" ISSUE ", text)
  
  # remove html tags
  # text = gensim.parsing.preprocessing.strip_tags(text)
  
  # remove punctuation
  text = text.translate(punctuations_trans)
  
  # remove numerics
  # text = gensim.parsing.preprocessing.strip_numeric(text)
  
  # remove non-ascii characters
  text = non_ascii_char_regex.sub("", text)
  
  text = unicodedata.normalize('NFD', text)
  
  # remove consecutive whitespace characters and convert tabs to spaces
  text = gensim.parsing.preprocessing.strip_multiple_whitespaces(text)
  
  # limit the number of tokens
  if max_tokens is not None:
    text = " ".join(text.split()[:max_tokens])
  
  return text

sample_df = sample_csv(train_file, 50_000)

q=[.5, .75, .8, .85, .9, .95, .99, .999]

display("title token count quantiles", sample_df["title"].apply(preprocess).apply(count_tokens).quantile(q=q))
display("body token count quantiles", sample_df["body"].apply(preprocess).apply(count_tokens).quantile(q=q))

# transform dataset into simpletransformers format
# https://simpletransformers.ai/docs/classification-data-formats/#multi-class-classification

def preprocess_row(row):
  doc = "TITLE " + preprocess(row["title"], max_tokens=20) # 99% of titles fit
  doc += " BODY " + preprocess(row["body"], max_tokens=511-count_tokens(doc))

  assert count_tokens(doc) <= 512

  return doc

def transform_to_simpletransformers_format(i_path, o_path):
	label_map = {"bug": 0, "feature": 1, "question": 2, "documentation": 3 }

	with open(i_path, "r") as i_f, open(o_path, "w") as o_f:
		reader = csv.DictReader(i_f)
		writer = csv.DictWriter(o_f, fieldnames=["text", "labels"], delimiter="\t")
		writer.writeheader()
		total = count_csv_rows(i_path)
		for row in tqdm(reader, desc="Transform to simpletransformers format", total=total):
			text = preprocess_row(row)
			labels = label_map[row["labels"]]
			writer.writerow({"text": text, "labels": labels})

transform_to_simpletransformers_format(train_file, "train.csv")
transform_to_simpletransformers_format(test_file, "test.csv")

print_csv_preview("train.csv", sep='\t')
print_csv_preview("test.csv", sep='\t')

### Training Loop


def model_args():
  timestamp = str(int(time.time()))

  # https://simpletransformers.ai/docs/usage/#configuring-a-simple-transformers-model
  args = ClassificationArgs()

  args.max_seq_length = 224
  args.learning_rate = 1e-5 # 4e-5
  args.num_train_epochs = 16
  args.train_batch_size = 64
  args.eval_batch_size = 64
  args.gradient_accumulation_steps = 1

  # custom evaluation metric
  # https://github.com/ThilinaRajapakse/simpletransformers/discussions/911
  args.use_early_stopping = False
  args.early_stopping_metric = "f1_micro"
  args.early_stopping_metric_minimize = False

  # evaluate at end of each epoch
  args.evaluate_during_training = True
  args.evaluate_during_training_steps = int(1e20) # never
  # args.evaluate_during_training_steps = 1.5  * 60 * 60 // args.gradient_accumulation_steps # target 1h

  # https://simpletransformers.ai/docs/classification-specifics/#lazy-loading-data
  args.lazy_loading = True

  args.save_steps = -1
  args.logging_steps = max(1, 1.6 * 30 // args.gradient_accumulation_steps) # target 30s
  args.manual_seed = 0

  args.output_dir = f"outputs/{timestamp}"
  args.best_model_dir = f"{args.output_dir}/best_model"

  # https://docs.wandb.ai/guides/integrations/other/simpletransformers
  # https://simpletransformers.ai/docs/tips-and-tricks/#visualization-support
  args.wandb_project = "NLBSE'23 Issue Report Classification"
  args.wandb_kwargs = {"entity": "nlbse", "notes": f"timestamp:{timestamp}"}

  return args

metrics = {
  "precision_bug": partial(sklearn.metrics.precision_score, average=None, labels=[0]),
  "recall_bug": partial(sklearn.metrics.recall_score, average=None, labels=[0]),
  "f1_bug": partial(sklearn.metrics.f1_score, average=None, labels=[0]),

  "precision_feature": partial(sklearn.metrics.precision_score, average=None, labels=[1]),
  "recall_feature": partial(sklearn.metrics.recall_score, average=None, labels=[1]),
  "f1_feature": partial(sklearn.metrics.f1_score, average=None, labels=[1]),

  "precision_question": partial(sklearn.metrics.precision_score, average=None, labels=[2]),
  "recall_question": partial(sklearn.metrics.recall_score, average=None, labels=[2]),
  "f1_question": partial(sklearn.metrics.f1_score, average=None, labels=[2]),

  "precision_documentation": partial(sklearn.metrics.precision_score, average=None, labels=[3]),
  "recall_documentation": partial(sklearn.metrics.recall_score, average=None, labels=[3]),
  "f1_documentation": partial(sklearn.metrics.f1_score, average=None, labels=[3]),

  "precision_micro": partial(sklearn.metrics.precision_score, average='micro'),
  "recall_micro": partial(sklearn.metrics.recall_score, average='micro'),
  "f1_micro": partial(sklearn.metrics.f1_score, average='micro'),
}

model = ClassificationModel(
  'roberta', 
  'roberta-base', 
  args=model_args(), 
  num_labels=4
)

model.train_model(train_df="train.csv", eval_df="test.csv", **metrics)

model = model = ClassificationModel(
	"roberta",
	"outputs/best_model",
  args=model_args(), 
  num_labels=4,
)

results, model_outputs, wrong_pred = model.eval_model(eval_df="test.csv", **metrics)
results

### Evaluation

In this section we evaluate our trained model.
We uploaded our trained model to the HuggingFace Hub. You can find it [here](https://huggingface.co/rafaelkallis/tickettagger-roberta).

from tqdm import tqdm
from transformers import pipeline
from datasets import load_dataset
from transformers.pipelines.pt_utils import KeyDataset
import sklearn.metrics

# classifier = pipeline(model="./best_model/", tokenizer="./best_model/")
classifier = pipeline(model="rafaelkallis/tickettagger-roberta", max_length=224, truncation=True, device=0, batch_size=256)

dataset = load_dataset('csv', data_files={"test": "test.csv"}, delimiter='\t')

y_true = dataset["test"]["labels"]
id2label = {0: "bug", 1: "feature", 2: "question", 3: "documentation"}
y_true = [id2label[y] for y in y_true]

y_pred = list(tqdm(classifier(KeyDataset(dataset["test"], "text")), total=len(y_true)))
y_pred = [p["label"] for p in y_pred]

for label in ["bug", "feature", "question", "documentation"]:
  P_c = sklearn.metrics.precision_score(y_true, y_pred, average=None, labels=[label])[0]
  R_c = sklearn.metrics.recall_score(y_true, y_pred, average=None, labels=[label])[0]
  F1_c = sklearn.metrics.f1_score(y_true, y_pred, average=None, labels=[label])[0]
  print(f"=*= {label} =*=")
  print(f"precision:\t{P_c:.4f}")
  print(f"recall:\t\t{R_c:.4f}")
  print(f"F1 score:\t{F1_c:.4f}")
  print()


P = sklearn.metrics.precision_score(y_true, y_pred, average='micro')
R = sklearn.metrics.recall_score(y_true, y_pred, average='micro')
F1 = sklearn.metrics.f1_score(y_true, y_pred, average='micro')

print("=*= micro averages =*=")
print(f"precision:\t{P:.4f}")
print(f"recall:\t\t{R:.4f}")
print(f"F1 score:\t{F1:.4f}")
